machine learning unit testing

+++ mathrun/src/main/java/com/ait/agilebuild/mathrun/model/QuestionDefinition.java
@@ -13,10 +13,9 @@ public class QuestionDefinition {
 	                  
 	private float[] attempts;
 	
-	private long[] attempt_duration;							    //time taken each attempt
+	private int[] attempt_duration;							    //time taken each attempt
 	
 	private int time;                                               //how long does kid spend on each question
-	
 	
 	public float getCorrect_answer() {
 		return correct_answer;
@@ -62,10 +61,10 @@ public class QuestionDefinition {
 	public void setAttempts(float[] attempts) {
 		this.attempts = attempts;
 	}
-	public long[] getAttempt_duration() {
+	public int[] getAttempt_duration() {
 		return attempt_duration;
 	}
-	public void setAttempt_duration(long[] attempt_duration) {
+	public void setAttempt_duration(int[] attempt_duration) {
 		this.attempt_duration = attempt_duration;
 	}
 	public int getTime() {
